Se modifica el servicio de registro de usuarios para la configuracion de los datos de inicio de sesion para el servicio de login

+++ src/main/java/com/school/app/repository/DireccionRepository.java
@@ -3,6 +3,6 @@ package com.school.app.repository;
 import com.school.app.entity.Direccion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DireccionRepository extends JpaRepository<Direccion, Integer> {
+public interface DireccionRepository extends JpaRepository<Direccion, Long> {
 
 }
